dev/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi>=23.1.0",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.22",
    "langchain-google-genai>=2.1.2",
    "langchain-openai>=0.3.11",
    "langchain-postgres>=0.0.13",
    "langgraph>=0.3.21",
    "langgraph-checkpoint-postgres>=2.0.19",
    "psycopg[binary,pool]>=3.2.6",
    "pydantic-settings>=2.8.1",
    "sqlalchemy[asyncio]>=2.0.40",
    "httpx[http2]>=0.28.1",
    "redis[hiredis]>=5.2.1",
    "cryptography>=44.0.2",
]

[dependency-groups]
dev = [
    "alembic>=1.15.2",
]
